pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rpg"
version = "0.1.0"
description = "Turn-based text RPG with deterministic world progression"
authors = [{name = "Your Team"}]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
	"sqlalchemy>=2.0.29",
	"mysql-connector-python>=8.3.0",
	"httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
	"pytest>=9.0.0",
	"pytest-cov>=5.0.0",
]
ui = [
	"rich>=13.7.0",
]
ops = [
	"python-dotenv>=1.0.1",
	"alembic>=1.13.1",
	"orjson>=3.10.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
